Read portfolio fills from the persistent history
usePortfolio now sources fills from IndexedDB via useTraderFills, keyed
by the connected trader, so history survives refresh and disconnect.
The mark price still reads the mock auction feed (market-wide state,
swapped by the tape work). CSV export is untouched — it was already a
pure function over Fill[].
